// bundles/nl.esi.comma.scenarios.dashboard/app/src/graph/types.d.ts
///
/// Copyright (c) 2024, 2025 TNO-ESI
///
/// See the NOTICE file(s) distributed with this work for additional
/// information regarding copyright ownership.
///
/// This program and the accompanying materials are made available
/// under the terms of the MIT License which is available at
/// https://opensource.org/licenses/MIT
///
/// SPDX-License-Identifier: MIT
///

export type EdgeType = 'cont' | 'curved' | 'bezier';

export interface GraphOptions {
    edgeType: EdgeType,
    view: 'interactive' | 'svg',
    spacing: number,
    minTestSet: number,
}